refactor: rename validation utils as helpers
